+++ app/src/main/kotlin/example/poeaa/project/controller/ProjectRequest.kt
@@ -1,4 +1,4 @@
-package example.poeaa.project
+package example.poeaa.project.controller
 
 data class ProjectRequest (
     var projectId: String? = null
